l - przed zmianami i czyszczeniem z nadmiernych elementów

+++ poligon inter/Model/BrokerDB.cs
@@ -1,14 +1,9 @@
 ﻿using Dapper;
-using MaterialDesignThemes.Wpf;
 using Microsoft.Data.Sqlite;
-using poligon_inter.ViewModel;
 using System.Collections.ObjectModel;
-using System.Data;
 using System.Diagnostics;
 using System.IO;
-using System.Windows;
 
-using String = System.String;
 
 namespace poligon_inter.Model;
 
@@ -21,44 +16,29 @@ namespace poligon_inter.Model;
 public class BrokerDB
 {
 
-    //private SqliteConnection s_conn;
-
-
     #region BrokerDB
-    // to pozostaje prywatne odwołąnie się do bazy danych powinno nastąpić poprzez
-    // podanie elementu lub nazwy bazy danych a nie do elementu con
-    // dostę do danych następuje poprzez podanie zaznaczonego elementu z którego jest wybierana baza
-    // w metodzie zwracającej określony zestaw danych
-    // Broker zawiera wszystkie niezbędne kody SQL, co umożliwia w razie potrzeby zmianę bazy
-    private Dictionary<string, SqliteConnection>? DB;
+
+    private Dictionary<string, SqliteConnection>? DB = null;
     //to aktualnie wybrana baza danych, wymagane jest sprawdzanie czy nie jest null
     private SqliteConnection? s_conn = null;
     public BrokerDB()
     {
-        // to będzie konstruktor
-        // ma przeszukać katalog z bazami i podłączyć bazy które tam znajdzie 
-        // do słownika
         string path = Tools.GetUserAppDataPath;
         DB = new();
         if (Directory.Exists(path))
         {
             string[] files = Directory.GetFiles(path, "*.db");
-            //SqliteConnection conn;
-            //string dbName = string.Empty;
-
             foreach (string file in files)
             {
-                //conn = new SqliteConnection( "Data Source=" + file);
-                //int start = file.LastIndexOf("\\") + 1;
-                //int len = file.Length - start - 3;
-                //dbName = file.Substring(start,len);
                 //tu powinno być sprawdzenie nazy bazy w bazie
                 //jak string.empty to dopiero kombinowanie z nazwy pliku
+                //lub nadawać jakiśunikaly identyfikator tak żeby mogły być dwie bazy o niby takich samych nazwach
+                // a urzytkownik nie musi o tym wiedzieć
+                // lub pozostawić rekord w tree do przechowywania nazwy bazy i metodę która to zwraca
                 var dbName = file.Substring(file.LastIndexOf("\\") + 1,
                     (file.Length - (file.LastIndexOf("\\") + 1)) - 3);
-                //Debug.WriteLine("dbname: " +file+ " , "+dbName);
-                DB[dbName] = new SqliteConnection("Data Source=" + file);
-                //DB[dbName].Open();
+            
+                DB[dbName] = new SqliteConnection("Data Source=" + file);                
             }
         }
     }
@@ -67,29 +47,21 @@ public class BrokerDB
 
     #endregion
 
-    //dotyczy wybranej bazy dabch, zastanawiam sięjak to rozwiązać, może zrobić metody statyczne ??
-    //cholera wie
+    // tu chodzi o nazwę bazy widzianą w drzewie a nie o nazwę pliku bazy
     private string DBName
     {
         get => s_conn != null ? s_conn.QuerySingle<string>(@"SELECT wartosc FROM slowniki Where nazwa = 'DBname'") : "";
         set
         {
+            //tu należy sprawdzić czy nie ma przypadkiem już rekordu o takiej nazwi 
+            //jak jest to robimy update a nie insert
             if ((s_conn != null) && (value != null))
                 s_conn.Execute(@"INSERT INTO slowniki (nazwa,wartosc) values ('DBname',@value)", new { value });
         }
     }
 
-    //broker
-    private void CreateDirX(string path)
-    {
-        //string dir = ;
-        //MessageBox.Show(dir);
-        Directory.CreateDirectory(path.Substring(0, path.LastIndexOf('\\')));
-    }
 
-
-
-    private void DbuildDb(string name = "")
+    private void DbuildTableDb(string name)
     {
         //var param = "jakaś baza";
         if (s_conn != null)
@@ -101,8 +73,10 @@ public class BrokerDB
             // słowniki to wartości różne przechowywane w jednej tabeli, ale nie jest ich aż tyle żeby tworzyć osobną tabele dla nich
             s_conn.Execute("CREATE TABLE IF NOT EXISTS slowniki (nazwa,wartosc , Id  INTEGER PRIMARY KEY)");
             //kategorie to foldery w drzewie, ple view odpowiada za widok jaki się wyświetli po kliknięciu LPM            
-            s_conn.Execute("CREATE TABLE IF NOT EXISTS Categories ( Name ,view , ParentID  INTEGER NOT NULL, Id    INTEGER NOT NULL UNIQUE, PRIMARY KEY(Id AUTOINCREMENT))");
+            s_conn.Execute("CREATE TABLE IF NOT EXISTS Categories ( Name ,View , ParentID  INTEGER NOT NULL, Id    INTEGER NOT NULL UNIQUE, PRIMARY KEY(Id AUTOINCREMENT))");
             s_conn.Execute("INSERT INTO slowniki (nazwa,wartosc) values ('DBname',@name)", new { name });
+            s_conn.Execute("INSERT INTO Categories (Name ,view , ParentID) values ('Pliki',@view,0)", new { view = "Files" });
+            s_conn.Execute("INSERT INTO slowniki (nazwa,wartosc) values ('extension',@name)", new { name = "*.avi,*.mp4,*.mkv" });
         }
     }
 
@@ -117,26 +91,52 @@ public class BrokerDB
                 {
                     s1 = s1.Parent;
                 }
-                return s1.Name;
+                return s1.GetKeyDB;
             }
-            else return SelectedItem.Name;
+            else return SelectedItem.GetKeyDB;
         }
         return string.Empty;
     }
 
-    public void AddDB(string name)
+    public TreeModel? CreateDB(string name)
     {
         string path = Tools.GetUserAppDataPath;
         if (DB == null) DB = new();
         if (!Directory.Exists(path)) Directory.CreateDirectory(path);
+        //dodany do słownika
         DB[name] = new SqliteConnection("Data Source=" + path + "\\" + name + ".db");
         s_conn = DB[name];
-        DbuildDb(name);
-    }
-    public int AddCategory(string category, TreeModel<Guid> SelectedItem)
-    {   //brak sprawdzania czy w głównym katalogu nie ma już takiej samej bazy
+        //budowa struktury
+        DbuildTableDb(name);
+        return GetTreeDB(name);
 
-        //throw new Exception("nie wybrano bazy");
+    }
+
+    public void DeleteDB(TreeModel DBx)
+    {
+        //Debug.WriteLine("usuwamy bazę");
+        if (DB == null) return;
+        //Debug.WriteLine("usuwamy bazę 2");
+        try
+        {
+            s_conn = DB[DBx.GetKeyDB];
+            s_conn.Close();
+            s_conn.Dispose();
+            SqliteConnection.ClearAllPools();
+            var dsource = s_conn.DataSource;
+            s_conn = null;
+            File.Delete(dsource);
+            DB.Remove(DBx.GetKeyDB);
+        }
+        catch (Exception ex)
+        {
+            Debug.WriteLine(ex);//tu dodać wrzucanie do logów błędu
+        }
+    }
+
+    public int AddCategory(string category, TreeModel<Guid> SelectedItem)
+    {
+        if (DB == null) return -1;
         s_conn = DB[GetTreeDBName(SelectedItem)];
 
         s_conn.Execute("insert into Categories (Name,ParentID) values (@category,@id)",
@@ -151,51 +151,78 @@ public class BrokerDB
     {
 
         //throw new Exception("nie wybrano bazy");
-        s_conn = DB[GetTreeDBName(SelectedItem)];
-        s_conn.Execute("Delete From Categories WHERE id = @id", new { id = SelectedItem.Id });
-        s_conn = null;
+        if (DB == null) return;
+        if (SelectedItem.Parent != null)
+        {
+            s_conn = DB[GetTreeDBName(SelectedItem)];
+            s_conn.Execute("Delete From Categories WHERE id = @id", new { id = SelectedItem.Id });
+            s_conn = null;
+        }
+        else
+        {
+
+        }
     }
 
     #region Tree
+
+
     public ObservableCollection<TreeModel>? GetTree()
     {
-        ObservableCollection<TreeModel> Tree = new();
-        TreeModel branch = null;
+        ObservableCollection<TreeModel> Tree = new();        
+        if (DB == null) return null;
         foreach (KeyValuePair<string, SqliteConnection> con in DB)
         {
             if (con.Value != null)
             {
-                //con.Value.Open();
-                branch = new TreeModel { Name = con.Key, IsExpanded = true, Parent = null };
-                // zwraca tree a nie tree<Guid>, jakby się dało to przerobić  bo wkuża okropnie
-                s_conn = con.Value;
-                branch.Children = GetTreBranch();
-                if ((branch.Children != null) && (branch.Children.Count > 0))
-                    foreach (var child in branch.Children)
-                        child.Parent = branch;
-                Tree.Add(branch);
-                if (con.Value == null) Debug.WriteLine("GetTreBranch is null");
+                Tree.Add(GetTreeDB(con.Key));
+                Debug.WriteLine("Key: "+con.Key+" Value: "+con.Value);
             }
         }
         s_conn = null;
         return Tree;
     }
 
-    public ObservableCollection<TreeModel<Guid>>? GetTreBranch()
+    /// <summary>
+    /// zwraca drzewo z bazy o podanej nazwie o ile ta baza jest dołączona do słownika DB
+    /// </summary>
+    /// <param name="name">nazwa bazy</param>
+    /// <returns>TreeModel - w raz z pod gałęziami</returns>
+    private TreeModel? GetTreeDB(string name)
+    {        
+        TreeModel? branch = null;        
+        
+        if ((DB != null)&&(DB[name] != null))
+        {   
+            s_conn = DB[name];
+            string ViewName = this.DBName;
+            branch = new TreeModel { Name = ViewName, IsExpanded = true, Parent = null };
+            branch.Children = GetTreeBranch();
+            branch.GetKeyDB = name;
+            if ((branch.Children != null) && (branch.Children.Count > 0))
+                foreach (var child in branch.Children)
+                    child.Parent = branch;
+        }
+        s_conn = null;
+        return branch;
+    }
+
+    // przejżeć obie matody naspokojnie i zrobić z nich jedną, są podobne
+    private ObservableCollection<TreeModel<Guid>>? GetTreeBranch()
     {
-        //_con można zrobić pole statyczne do którego będzie przypisana aktualnie wybrana baza
-        // unikamy w ten sposób przekazywania tej bazy między metodami
-        ObservableCollection<TreeModel<Guid>> Tree = null;
+        ObservableCollection<TreeModel<Guid>>? Tree = null;
         //to i tak jest na początek później w planach jest zmiana tego na jakiś obiekt
         if (s_conn != null)
+        {
             Tree = new ObservableCollection<TreeModel<Guid>>(
             s_conn.Query<TreeModel<Guid>>("SELECT * FROM Categories WHERE ParentID == 0", new DynamicParameters()));
-        foreach (var item in Tree)
-        {
-            //item.IsExpanded = false;
-            item.IsExpanded = true;
-            item.IsSelected = false;
-            LoadSubCategories(item);
+            foreach (var item in Tree)
+            {
+                //item.IsExpanded = false;
+                item.IsExpanded = true;
+                item.IsSelected = false;
+                LoadSubCategories(item);
+            }
         }
         return Tree;
     }
@@ -219,22 +246,23 @@ public class BrokerDB
     */
     private void LoadSubCategories(TreeModel<Guid> item)
     {
-        // to jest dzięki jakiemuś frameworkowi ale nie wiem jakiemu
-        //item.SubCategories = new ObservableCollection<Category>(s_conn.Category.Where(x => x.ParentID == item.Id));
-        //to zamienić na uruchomioną  bazę !!!
-        item.Children = new ObservableCollection<TreeModel<Guid>>(
-        s_conn.Query<TreeModel<Guid>>("select * from Categories where ParentID == @Id", item));//!!!
-                                                                                               //item.AddChild(sub);
-        foreach (var subitem in item.Children)
+        if (s_conn != null)
         {
-            subitem.Parent = item;
-            subitem.IsExpanded = true;
-            subitem.IsSelected = false;
-            LoadSubCategories(subitem);
+            item.Children = new ObservableCollection<TreeModel<Guid>>(
+            s_conn.Query<TreeModel<Guid>>("select * from Categories where ParentID == @Id", new { Id = item.Id }));//!!!
+                                                                                                                //item.AddChild(sub);
+            foreach (var subitem in item.Children)
+            {
+                subitem.Parent = item;
+                subitem.IsExpanded = true;
+                subitem.IsSelected = false;
+                LoadSubCategories(subitem);
+            }
         }
     }
 
     #endregion Tree
+    /*
     public void AddPassword(string password)
     {
         string pass = Tools.CalculateMD5Sting(password);
@@ -242,11 +270,12 @@ public class BrokerDB
         command.CommandText = String.Format(@"insert into slowniki (nazwa,wartosc) values (pass,'{0}')", pass);
         command.ExecuteNonQuery();
     }
-
+    */
     #region Files
     public void addFile(string nazwa, string path, string MD5 = "")
     {
         //metoda do przepisania na nowo
+        /*
         string id = string.Empty;
         var command = s_conn.CreateCommand();
         command.CommandText = String.Format("SELECT id FROM katalogi WHERE path='{0}'", path);
@@ -275,6 +304,7 @@ public class BrokerDB
         command.CommandText = String.Format(@"insert into pliki (nazwa,id_katalogu,usuniety,rozszezenie,MD5) values ('{0}','{1}',false,'','{2}')", nazwa, id, MD5);
         var ile = command.ExecuteNonQuery();
         if (ile < 1) MessageBox.Show(" chyba coś poszło nie tak, ilość dodana: " + ile);
+        */
     }
 
     #endregion Files
@@ -282,4 +312,3 @@ public class BrokerDB
 
 
 }
-
